Updated zero-shot predictions
label_counter.py uses clip to encode the labels in word_labels.yaml and the images in the directory you provide to compute the similarity percentage between each text-image pair. It then uses the labels in word_labels.yaml as keys for a Counter. The top 5 similarity percentages for each image are updated in the counter (so each word in the counter is followed by the sum of its corresponding percentage when it scored in the top 5 similarity labels for each image). This counter is then dumped and saved to a pickle file called label_counter.pkl. This program also creates a dictionary containing each word in word_labels.yaml as keys and a list of the corresponding similarity percentages as values. For example, the list that corresponds to the key "boat" contains all the similarity percentages when the label "boat" is one of the top 5 matches for an image. This dictionary is dumped and saved to a pickle file called similarity_dict.pkl.

I have a directory of approximately 154,000 images, and I put 8 labels into word_labels.yaml, and this program (connected to CUDA) ran in about 40 mins. Pickling the data in the counter and the dictionary allows me to save/access/analyze them without having to rerun label_counter.py.