Update predictions_hist.py
predictions_hist.py loads the dictionary from similarity_dict.pkl so that it can be used to create and save multiple histograms. One histogram is made for each word in word_labels.yaml. The histogram plots the distribution of similarity percentages that are greater than or equal to the probability specified in word_labels.yaml. The plot title contains the count of files represented by the plot. This program makes a directory called "histograms" where each plot is saved.