Keyword search program
keyword_search.py reads faiss/keyword_search.yaml to get a path to an image dataset, a path to an index file, a list of keywords to search for, and k (the number of nearest neighbors to find). The program uses CLIP to tokenize each word in the keywords list and, using the index, finds finds the k nearest neighbor images to each word. The images are printed by the program and are copied into a directory named after its corresponding word. These image directories are in a directory called knn.